refactor: isolate contracts and implementatios for repositories
