+++ src/domain/services/user/delete-user-profile.service.spec.ts
@@ -3,8 +3,8 @@ import { DeleteUserProfileService } from './delete-user-profile.service';
 import { nonExistentUserId } from '@/core/shared/utils';
 import { NotFoundException } from '@nestjs/common';
 import { Test, type TestingModule } from '@nestjs/testing';
-import { UserRepository } from '@/infra/repositories/user.repository';
-import { PrismaUserRepository } from '@/infra/repositories/prisma/prisma-user-repository';
+import { UserRepository } from '@/core/repositories/user.repository';
+import { PrismaUserRepository } from '@/domain/repositories/prisma-user-repository';
 import { PrismaService } from '@/infra/database/prisma.service';
 import { CreateMockUser } from '@/test/mocks/create-mock-user/create-mock-user';
