+++ src/domain/services/user/get-user-profile.service.spec.ts
@@ -3,8 +3,8 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { nonExistentUserId } from '@/core/shared/utils';
 import { GetUserProfileService } from '@/domain/services/user/get-user-profile.service';
 import { PrismaService } from '@/infra/database/prisma.service';
-import { PrismaUserRepository } from '@/infra/repositories/prisma/prisma-user-repository';
-import { UserRepository } from '@/infra/repositories/user.repository';
+import { PrismaUserRepository } from '@/domain/repositories/prisma-user-repository';
+import { UserRepository } from '@/core/repositories/user.repository';
 import { MockPrismaService } from '@/test/mocks/prisma';
 import { CreateMockUser } from '@/test/mocks/create-mock-user/create-mock-user';
